Added RangedList class for tracking bytes in files
This won't matter for well-formed TIFFs, but sometimes bytes can
overlap, so it's a good idea to keep track of which regions belong to
what things within the file, in case something tries to reuse an offset
for a different purpose.

This information could help a lot toward remediation in particular.