Consolidated 3to2 code
There are a few things that 3to2 doesn't translate very well (mostly
bytes stuff), so I put that all at the beginning of the file. This way,
it'll be a little faster whenever I'm generating python2.7 versions of
stuff.

Beyond that, there's a zero comparison that should be "\0" in 2.7, and
unittests need assertRaisesRegexp instead of assertRaisesRegex. Also,
2.7 could give longs instead of ints.