Wrote a lot of pdf reader code
It's now able to read the entire xref history. It's also able to deal
with simple indirect objects. It still needs to be able to deal with
streams (compressed streams in particular).

I suppose the next step is to design compression helper functions and to
also point out where decompression (and compression) are beyond the
scope of what we'd even want to do here.

After that, it's back to object streams. I'll also need to add code for
cross reference streams.