Handled with argparse now
Previously, bin/deep-blue was handed a list of paths to volumes ready to
be converted to pdf. It had a static output path, and that was just
silly. Now it takes that path as an argument.

Also, by default, it doesn't overwrite any existing pdfs. With the `-f`
argument, it can be forced to output pdfs regardless of what's already
there.